requirements.txt
Flask==2.0.1
numpy==1.19.5
Pillow==8.3.2
requests==2.25.1